Created a graph to visualize the decisions.
Current version uses a sample of only 200 contracts, but uses all of the copies of the snapshots that we currently have.
SQL code updated to be able to produce the decisions table needed, plus used it to generate a sample data set and a the .html file produced. These are the outputs of the visualize_decisions.py and decisions_table.sql.